create book Module
